Add a context listener that emulates classpath scanning
The context listener is explicitly defined in web.xml so that it is
discoverd and used without classpath scanning.

Manually instantiates the relevant ServletContainerInitializer instances
and provides each of them a set of classes that it's interested in.

After that, run Vaadin's own context listener that deploys a
preconfigured servlet.